restructure - combine significant and random experiment files into one for globals to work
